Move config to package
