+++ internal/core/core.go
@@ -19,13 +19,12 @@ package core
 import (
 	"context"
 	"fmt"
+	"github.com/media-streaming-mesh/msm-cp/pkg/config"
 	node_mapper "github.com/media-streaming-mesh/msm-cp/pkg/node-mapper"
 	"github.com/media-streaming-mesh/msm-cp/pkg/transport"
 	"net"
 	"os/signal"
 	"syscall"
-
-	"github.com/media-streaming-mesh/msm-cp/internal/config"
 )
 
 // App contains minimal list of dependencies to be able to start an application.
