+++ internal/core/wire.go
@@ -21,8 +21,8 @@ package core
 
 import (
 	"github.com/google/wire"
-	"github.com/media-streaming-mesh/msm-cp/internal/config"
 	"github.com/media-streaming-mesh/msm-cp/internal/rtm"
+	"github.com/media-streaming-mesh/msm-cp/pkg/config"
 )
 
 func InitApp() (*App, error) {
